Iterative update
I could go live with this.
The fa-newspaper icon is not working for some reason so I wrote the dev; it's the only one I can't get to work of course.
I'd like to add favicons just not sure if I want to dump in root (I should that's so easy; but I won't)
Last but not least, I want a more original background

Signed-off-by: Chris Short <chris@chrisshort.net>